Modified GUI, added Home Functionality, minor error fixes
Condensed user text input into a single input, added Home and Sethome functions, Optimized movement part of the code so that it only runs when the target position is updated.

NEEDED CHANGES
Modify movement code because currently it just repeats the same movement backwards after a 5sec delay. Target moving to position and staying there, only moving back if a new position is inputted or if home is inputted. Remove duplicate movement for loops by creating a single function to stream line code.